requirements.txt
streamlit>=1.30.0
langgraph>=0.0.26
langchain-core>=0.1.52
langchain-google-genai>=1.0.3
python-dotenv>=1.0.0